third_party/WebKit/LayoutTests/external/wpt/background-fetch/fetch.https.window-expected.txt

This is a testharness.js-based test.
PASS Background Fetch requires an activated Service Worker
PASS Argument verification is done for BackgroundFetchManager.fetch()
PASS IDs must be unique among active Background Fetch registrations
PASS Using Background Fetch to successfully fetch a single resource
PASS Background Fetch that exceeds the quota throws a QuotaExceededError
FAIL Fetches can have requests with duplicate URLs promise_test: Unhandled rejection with value: object "TypeError: Fetches with duplicate requests are not yet supported. Consider adding query params to make the requests unique. For updates check http://crbug.com/871174"
FAIL Fetches can have requests with a body promise_test: Unhandled rejection with value: object "TypeError: Requests with a body are not yet supported. For updates check http://crbug.com/774054"
PASS recordsAvailable is false after onbackgroundfetchsuccess finishes execution.
Harness: the test ran to completion.
